Fixed: failing decoding json with incorrect type
Sometimes 'v' parameter is double

+++ app/src/main/java/com/khoben/ticker/api/finnhub/http/model/CandleStockResult.kt
@@ -18,5 +18,5 @@ data class CandleStockResult(
     @SerialName("t")
     val timestamps: List<Int>? = null,
     @SerialName("v")
-    val volumes: List<Int>? = null
+    val volumes: List<Double>? = null
 )
